+++ src/Queue/src/Job/CallableJob.php
@@ -15,7 +15,7 @@ final class CallableJob implements HandlerInterface
     ) {
     }
 
-    public function handle(string $name, string $id, array $payload, array $context = []): void
+    public function handle(string $name, string $id, array $payload, array $headers = []): void
     {
         if (!isset($payload['callback'])) {
             throw new InvalidArgumentException('Payload `callback` key is required.');
@@ -30,7 +30,7 @@ final class CallableJob implements HandlerInterface
             [
                 'name' => $name,
                 'id' => $id,
-                'context' => $context,
+                'headers' => $headers,
             ]
         );
     }